refactor : 패키지 분리 및 메서드 네이밍 개선
