+++ docs/README.md
@@ -9,7 +9,7 @@
   - [x] 단, mp를 넘어서는 공격을 할 수는 없다.
 - [x] 2: 몬스터 랜덤 공격 생성
 - [x] 3: 각 공격 적용
-  - [x] 단, 최대치?✅ 이상의 MP는 회복할 수 없다.
+  - [x] 단, 최대치?  이상의 MP는 회복할 수 없다.
   - [x] 단, 몬스터, 플레이어의 남은 Hp가 minus일 수는 없다. -> 0으로 통일 출력
   - [x] 게임 횟수도  올라간다.
 - [x] 4: 공격 적용후, 몬스터, 유저 상태 출력
@@ -90,14 +90,14 @@
 ### 요구사항
 - [x] indent (2)
 - [x] 메서드 길이 (15)
-- [ ] 단위 테스트 확인 
+- [x] 단위 테스트 확인 
 ### 그 외
 - [x] 컨트롤러 메소드 네이밍 정리 및 분리 개선
 - [x] 숫자 리터럴 상수화 (outputFormatter, inputView, inputValidator도)
 ---
 - [x] 부생성자가 있는 경우 생성자의 접근제어자가 private인지 
-- [ ] 이해를 위한 코드 전반의 공백 추가
-- [ ] 각 클래스 메소드 순서 점검하기
-- [ ] 테스트 코드 리팩토링
-- [ ] getter 리스트 레퍼런스에 unmodifaible 추가
+- [x] 이해를 위한 코드 전반의 공백 추가
+- [x] 각 클래스 메소드 순서 점검하기
+- [x] 테스트 코드 리팩토링
+- [x] getter 리스트 레퍼런스에 unmodifaible 추가
 ----
